% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{min_records}
\alias{min_records}
\title{Demand minimal number of records}
\usage{
min_records(n)
}
\arguments{
\item{n}{Minimal number of records in a group.}
}
\value{
a function that accepts a data frame and returns \code{TRUE}
        when the number of records is larger than or equal to \code{n}
        and otherwise \code{FALSE}.
}
\description{
Demand minimal number of records
}
\examples{

min_records(5)(women)
min_records(200)(women)

}
\seealso{
Other helpers: 
\code{\link{frac_complete}()},
\code{\link{min_complete}()}
}
\concept{helpers}
