% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{object_list}
\alias{object_list}
\alias{format.object_list}
\alias{print.object_list}
\alias{[.object_list}
\title{Create a classed list}
\usage{
object_list(x)

\method{format}{object_list}(x, ...)

\method{print}{object_list}(x, ...)

\method{[}{object_list}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a list}
}
\description{
Classed lists are used to pretty-print a list that is stored
in a data frame.
}
\examples{
object_list(list(lm(speed ~ dist, data=cars)))

}
\keyword{internal}
